Revert "Refactored and updated NGL dep"
This reverts commit 3d98817be3dfa702b9d8b54ac3bbc14df6927d80.
