+++ js/containers/tindspectRoot.js
@@ -4,17 +4,17 @@
 import React, { Component } from 'react'
 import { Provider } from 'react-redux'
 import configureStore from '../configureStore'
-import App from './app';
+import Tindspect from '../components/tindspectHolder';
+
 
 const store = configureStore()
  
-export default class Root extends Component {
+export default class PickerRoot extends Component {
     render() {
-        
         return (
             <Provider store={store}>
-                <App></App>
-            </Provider>
+            <Tindspect></Tindspect></Provider>
         )
     }
 }
+
