+++ js/index.js
@@ -1,10 +1,10 @@
 import 'babel-polyfill'
 import React from 'react'
 import { render } from 'react-dom'
-import Root from './containers/root'
+import PickerRoot from './containers/pickerRoot'
 import '../css/index.css';
 
 render(
-  <Root />,
+  <PickerRoot />,
   document.getElementById('app')
 )
