+++ package.json
@@ -30,7 +30,7 @@
     "css-loader": "^0.28.10",
     "jquery": "^3.3.1",
     "lodash": "^4.17.5",
-    "ngl": "2.0.0-dev.32",
+    "ngl": "2.0.0-dev.29",
     "nprogress": "^0.2.0",
     "react": "^16.2.0",
     "react-bootstrap": "^0.32.1",
